pyproject.toml
[project]
name = "agente-tiktok"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "agno==1.5.0",
    "python-dotenv>=1.2.2",
    "groq>=0.13.0",
    "fastapi>=0.135.1",
    "uvicorn>=0.38.0",
    "openai>=2.24.0",
    "sqlalchemy>=2.0.48",
    "tavily-python>=0.7.22",
]
